Prune and clean project dependencies and import style
Each subdirectory now contains a base file (e.g., `core/` contains a
file called `core.jl`).  This is the file we `include` in the module
file, `TextAssociations.jl`.  This allows us only one `include` for
each "submodule."

In addition to this, we prune unused project dependencies from the
project requirements.

Finally, we add compat entries for all project dependencies, in the
interest of best practice (and to keep CompatHelper quiet!).  This
means that this commit:

Closes #50 (Plots; removed)
Closes #51 (XLSX; compat=0.10)
Closes #52 (DataStructures; compat=0.18, 0.19 [requires 0.18 for
            TextAnalysis])
Closes #53 (Statistics; compat=1)
Closes #54 (SQLite; removed)
Closes #55 (Chain; compat=1)
Closes #56 (HypothesisTests; compat=0.11)
Closes #57 (FreqTables; compat=1)
Closes #58 (Glob; compat=1)
Closes #59 (ProgressMeter; compat=1)
Closes #60 (JSON; compat=0.21)
Closes #61 (HTTP; removed)